requirements.txt
numpy==1.20.2
opencv-python==4.5.4.60
opencv_contrib_python==4.5.4.60
opencv-python-headless==4.5.4.60
PyYAML==6.0
thop==0.0.31.post2005241907
torch==1.8.0
torchvision==0.9.1
